move gql helper types and add missing param
