+++ packages/spruce-skill-server/interfaces/ctx.ts
@@ -6,12 +6,8 @@ import { ISpruceServices } from './services'
 import { ISpruceUtilities } from './utilities'
 import { ISpruceAuth } from './auth'
 import { Sequelize } from 'sequelize/types'
-import SpruceCoreModel from '../lib/SpruceModel'
 import { ISpruceGQLTypes } from './gql'
-import { FindOptions } from 'sequelize'
-import { GraphQLObjectType } from 'graphql'
-
-type SpruceCoreModelType = typeof SpruceCoreModel
+import { ISpruceGQLHelpers } from '../gql/helpers'
 
 export interface ISpruceContext<
 	ISkillModels = ISpruceModels,
@@ -30,30 +26,7 @@ export interface ISpruceContext<
 	utilities: ISkillUtilities
 	gql: {
 		types: ISkillGQLTypes
-		helpers: {
-			attributes(model: SpruceCoreModelType, options?: Record<string, any>): any
-			buildConnection(options: {
-				/** Name your connection. This must be unique */
-				name?: string
-				/** The model you're connecting. For example ctx.db.models.User */
-				model: SpruceCoreModelType
-				associationName: string
-				type: GraphQLObjectType
-				connectionOptions: {
-					before?: (
-						findOptions: FindOptions,
-						args: Record<string, any>,
-						context: any
-					) => Promise<FindOptions>
-					// TODO: Define after
-					// after?: (
-					// 	findOptions: FindOptions,
-					// 	args: Record<string, any>,
-					// 	context: any
-					// ) => Promise<FindOptions>
-				}
-			}): any
-		}
+		helpers: ISpruceGQLHelpers
 	}
 }
